feat(agent-tars-web-ui): refine replay control component
